% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{calculate_edge_density}
\alias{calculate_edge_density}
\title{Title}
\usage{
calculate_edge_density(adj_mat, clique_idx)
}
\arguments{
\item{adj_mat}{ajdacency matrix}

\item{alpha}{minimum density of clique}
}
\value{
indices corresponding to maximum partial clique of given adjacency matrix
}
\description{
Title
}
