% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_alpha.R
\name{compare_alpha}
\alias{compare_alpha}
\title{compare_alpha}
\usage{
compare_alpha(alpha, trial)
}
\arguments{
\item{alpha}{vector of values for alpha}

\item{trial}{number of trials}
}
\value{
dataset with results to know whether or not the provided clique_idx forms a valid partial clique.
}
\description{
compare_alpha
}
