% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique_master.R
\name{compute_correct_density}
\alias{compute_correct_density}
\title{Compute the correct density, given a set of nodes}
\usage{
compute_correct_density(adj_mat, clique_idx)
}
\arguments{
\item{adj_mat}{a symmetric \code{matrix} with only values \code{0} or \code{1}, has \code{1}'s along its diagonal, has no row- or column-names, and will have between 5 to 50  rows/columns (inclusive).}

\item{clique_idx}{a \code{numeric} vector of index numbers corresponding to the nodes (i.e., values between 1 and \code{nrow(adj_mat)}) that your function deems to be in the maximum partial clique. This vector cannot have duplicate elements, must be positive integers, and the largest value cannot exceed \code{nrow(adj_mat)}}
}
\value{
a numeric, which is the density of edges among \code{adj_mat[clique_idx,clique_idx]}
}
\description{
We will define a \code{clique_idx} of length 1 to have a density of 1.
}
