% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique13.R
\name{compute_maximal_partial_clique13}
\alias{compute_maximal_partial_clique13}
\title{Compute maximal partial clique}
\usage{
compute_maximal_partial_clique13(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{adj_mat Numeric matrix (n_n) of 0/1's, symmetric, 1's on diagonal,
no row/col names, with 5:50.}

\item{alpha}{Single numeric between 0.5 and 1, the minimum required edge-density.}
}
\value{
A list with clique_idx and edge_density
}
\description{
Compute maximal partial clique
}
