% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique14.R
\name{compute_maximal_partial_clique14}
\alias{compute_maximal_partial_clique14}
\title{Compute maximal partial clique function}
\usage{
compute_maximal_partial_clique14(adj_mat, alpha, top_k_seeds = NULL)
}
\arguments{
\item{adj_mat}{A nxn undirected graph with 1 meaning connected and 0 meaning unconnected. It is symmetric and should be generated by some generate_partial_clique_function. n is from 5 to 50 inclusive.}

\item{alpha}{The threshold of minimum edge density of the potential clique. It should range from 0.5 to 1 inclusive.}

\item{top_k_seeds}{A value that make algorithm more efficient by setting starter nodes with most connections. It is set to sqrt of number of nodes in matrix.}
}
\value{
A list with two elements:
\describe{
\item{clique_idx}{a numeric vector of node indices in the partial clique.}
\item{edge_density}{the edge density among those nodes.}
}
}
\description{
This function identifies a subset of nodes in a binary symmetric adjacency matrix
that form a maximal partial clique with edge density at least \code{alpha}.
It uses a greedy expansion algorithm starting from the top-k highest-degree nodes
and iteratively adds nodes while maintaining the required edge density. However,
is not guarantee the best solution in every task, limited by top-k, especially when
top-k is large or matrix size is large.
}
