% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique15.R
\name{compute_maximal_partial_clique15}
\alias{compute_maximal_partial_clique15}
\title{Find hte matximal partial clique}
\usage{
compute_maximal_partial_clique15(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{the graph}

\item{alpha}{a minimum threshold on when to stop}
}
\value{
list
}
\description{
Start from the first entry, and iteratively add more nodes.
The first moment you do not find a valid click, stop
}
