% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique16.R
\name{compute_maximal_partial_clique16}
\alias{compute_maximal_partial_clique16}
\title{Find maximal partial clique}
\usage{
compute_maximal_partial_clique16(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{matrix}

\item{alpha}{value}
}
\value{
list
}
\description{
Find maximal partial clique
}
