% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique18.R
\name{compute_maximal_partial_clique18}
\alias{compute_maximal_partial_clique18}
\title{A function to compute the size of the maximal partial clique of an adjacency matrix}
\usage{
compute_maximal_partial_clique18(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{A symmetric and diagonal adjacency matrix}

\item{alpha}{The threshold for the maximum number of edges in the partial clique}
}
\value{
An index vector showing the first maximal partial clique that exceeds the threshold, where the "first" is decided by the order of possible cliques determined by the \code{utils::combn} function. Also returns the edge density of that particular partial clique
}
\description{
Computes the maximum size of a partial clique in an adjacency matrix
}
