% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique20.R
\name{compute_maximal_partial_clique20}
\alias{compute_maximal_partial_clique20}
\title{Compute the Maximal Partial Clique}
\usage{
compute_maximal_partial_clique20(adj_mat, alpha, verbose = 0)
}
\arguments{
\item{adj_mat}{A square symmetric adjacency matrix where each element is either 0 or 1, and the diagonal elements are 0.}

\item{alpha}{A numeric value between 0.5 and 1 that specifies the minimum required edge density for the partial clique.}

\item{verbose}{boolean}
}
\value{
A list containing:
\itemize{
\item{clique_idx}{A numeric vector of indices representing the nodes in the maximal partial clique.}
\item{edge_density}{The edge density of the found clique.}
}
}
\description{
This function calculates the largest subset of nodes within an adjacency matrix that forms a partial clique meeting a specified edge density threshold. The function employs a heuristic approach to handle the computationally challenging nature of the task.
}
