% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique21.R
\name{compute_maximal_partial_clique21}
\alias{compute_maximal_partial_clique21}
\title{Find the maximal partial clique}
\usage{
compute_maximal_partial_clique21(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{an adjacency matrix with no row- or column-names,
between 5 to 50 rows/columns (inclusive).}

\item{alpha}{a value between 0.5 and 1 (inclusive), edge density}
}
\value{
a list with element \code{clique_ind}, a vector of index numbers of the largest partial clique,
and element \code{edge_density}, percentage of edges in \code{adj_mat} among the nodes in \code{clique_idx}
}
\description{
Computes the largest partial clique given an adjacency matrix \code{adj_mat}
and a required edge density \code{alpha}. This method computes the edge density of the partial clique
over all possible node combinations and selects out the one with the largest edge density.
}
