% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique22.R
\name{compute_maximal_partial_clique22}
\alias{compute_maximal_partial_clique22}
\title{Compute the Maximal Partial Clique}
\usage{
compute_maximal_partial_clique22(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{A symmetric matrix representing the adjacency matrix of a graph.}

\item{alpha}{A numeric value between 0.5 and 1 indicating the minimum edge density.}
}
\value{
A list with the indices of the nodes and the edge density of the clique.
}
\description{
This function identifies the largest subset of nodes in an adjacency matrix
that forms a partial clique meeting a specified minimum edge density (\code{alpha}).
}
