% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique23.R
\name{compute_maximal_partial_clique23}
\alias{compute_maximal_partial_clique23}
\title{Compute the largest partial clique}
\usage{
compute_maximal_partial_clique23(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{The adjacency matrix representing the graph.}

\item{alpha}{The required edge density for the partial clique, must be between 0 and 1.}
}
\value{
A list containing:
\item{node_ids}{A vector containing the indices of the vertices in the computed maximal partial clique.}
\item{edge_density}{The edge density of the computed maximal partial clique.}
}
\description{
This function computes the largest partial clique given an adjacency matrix and a required edge density alpha.
}
