% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique24.R
\name{compute_maximal_partial_clique24}
\alias{compute_maximal_partial_clique24}
\title{I'm not sure if this will work...}
\usage{
compute_maximal_partial_clique24(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{a matrix}

\item{alpha}{a value}
}
\value{
a list
}
\description{
This function will only output one edge, which will satisfy the definition
of a clique but will likely not be the largest one
}
