% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique25.R
\name{compute_maximal_partial_clique25}
\alias{compute_maximal_partial_clique25}
\title{This function identifies the largest partial clique based on the given adjacency matrix
and a required edge density.}
\usage{
compute_maximal_partial_clique25(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{A symmetric matrix (0-1) representing the adjacency relationships among nodes.}

\item{alpha}{A numeric value between 0.5 and 1 (inclusive) representing the required edge density.}
}
\value{
A list containing the index vector of nodes in the maximal partial clique (clique_idx)
and the computed edge density (edge_density).
}
\description{
This function identifies the largest partial clique based on the given adjacency matrix
and a required edge density.
}
