% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique3.R
\name{compute_maximal_partial_clique3}
\alias{compute_maximal_partial_clique3}
\title{Compute maximum partial clique.}
\usage{
compute_maximal_partial_clique3(adj_mat, alpha, verbose = 0)
}
\arguments{
\item{adj_mat}{Matrix. Adjacency matrix to search.}

\item{alpha}{Numeric. Minimum edge density of any potential partial clique.}

\item{verbose}{boolean}
}
\value{
List.
\item{clique_idx}{Nodes in the found maximum partial clique.}
\item{edge_density}{Edge density of found clique. Should be above alpha.}
}
\description{
Compute maximum partial clique.
}
