% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique4.R
\name{compute_maximal_partial_clique4}
\alias{compute_maximal_partial_clique4}
\title{Compute Maximal Partial Clique}
\usage{
compute_maximal_partial_clique4(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{A symmetric adjacency matrix (0/1) with diagonal 1s.}

\item{alpha}{A numeric value between 0.5 and 1, representing the minimum edge density required.}
}
\value{
A list with the following elements:
\describe{
\item{clique_idx}{A numeric vector of node indices that form the partial clique.}
\item{edge_density}{The edge density among the selected nodes (between 0 and 1).}
}
}
\description{
This function computes an approximate maximal partial clique from a given adjacency matrix,
such that the edge density among the selected nodes is at least \code{alpha}.
}
