% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique7.R
\name{compute_maximal_partial_clique7}
\alias{compute_maximal_partial_clique7}
\title{Find the maximal partial clique in an adjacency matrix}
\usage{
compute_maximal_partial_clique7(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{a symmetric matrix with 5 to 50 rows}

\item{alpha}{a numeric value between 0.5 and 1}
}
\value{
a list containing clique_idx and edge density
}
\description{
Find the maximal partial clique in an adjacency matrix
}
