% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique_9.R
\name{compute_maximal_partial_clique_9}
\alias{compute_maximal_partial_clique_9}
\title{Compute the maximal partial clique}
\usage{
compute_maximal_partial_clique_9(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{symmetric adjacency matrix with 5 to 50 rows/columns}

\item{alpha}{minimum edge density required for a valid partial clique}
}
\value{
A list containing a maximal partial clique, along with its edge density.
}
\description{
A function that computes the maximal partial clique within an adjacency
matrix given a set edge density for what constitutes a valid partial clique.
}
\details{
This function identifies the maximal partial clique within
an adjacency matrix and its edge density. It uses a greedy process,
starting with each node in the matrix and iteratively expanding to new
neighbors with the greatest number of edge connections.

If there are multiple maximal partial cliques found,
then only the first one identified is returned.
}
\author{
Nicos E. Soares \cr Maintainer: Nicos E. Soares \href{mailto:nicos.soares20@gmail.com}{nicos.soares20@gmail.com}
}
