% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique_john.R
\name{compute_maximal_partial_clique_master_safe}
\alias{compute_maximal_partial_clique_master_safe}
\title{Safe function of master function}
\usage{
compute_maximal_partial_clique_master_safe(
  adj_mat,
  alpha,
  number,
  time_limit = 30
)
}
\arguments{
\item{adj_mat}{a symmetric \code{matrix} of 0/1 with 1’s on the diagonal,}

\item{alpha}{a single \code{numeric} in [0.5, 1].}

\item{number}{an integer from 1 to 15 selecting the implementation.}

\item{time_limit}{maximum seconds before timing out.}
}
\value{
a list with \code{clique_idx}, \code{edge_density} (which are from the implementation itself)
and \code{status} and \code{valid} (which are added by this function)
}
\description{
Same functionality of master fucntion, only add a error catch for those
implementation who has error generate a matrix.
}
