% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique18.R
\name{.calculate_maximum_candidate_18}
\alias{.calculate_maximum_candidate_18}
\title{A function to compute the upper bound of the clique size}
\usage{
.calculate_maximum_candidate_18(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{A symmetric and diagonal adjacency matrix}

\item{alpha}{The threshold for the maximum number of edges in the partial clique}
}
\value{
The upper bound of the clique size
}
\description{
Given an adjacency matrix, this function uses the sum of the adjacency matrix to upper bound the possible clique size
}
