% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique18.R
\name{.fixed_combn_18}
\alias{.fixed_combn_18}
\title{Generate combinations while fixing the first k elements}
\usage{
.fixed_combn_18(vec, c, k)
}
\arguments{
\item{vec}{vector to choose from}

\item{c}{number of elements to choose}

\item{k}{number of elements to fix}
}
\value{
A matrix of combinations
}
\description{
Generate combinations while fixing the first k elements
}
