% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique18.R
\name{.minimum_searcher_18}
\alias{.minimum_searcher_18}
\title{A function that further restricts the candidate nodes for a partial clique}
\usage{
.minimum_searcher_18(adj_mat, candidates, c, at_least_total)
}
\arguments{
\item{adj_mat}{A symmetric and diagonal adjacency matrix}

\item{candidates}{A set of candidate nodes}

\item{c}{The size of the partial clique}

\item{at_least_total}{The minimum number of edges in the partial clique}
}
\value{
A minimum number of edges per node
}
\description{
Given a set of candidate nodes, this function further restricts the candidate nodes for a partial clique
}
