% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique18.R
\name{.pruned_cols_18}
\alias{.pruned_cols_18}
\title{Prunes columns that can be a part of a partial clique of size c}
\usage{
.pruned_cols_18(adj_mat, alpha, c)
}
\arguments{
\item{adj_mat}{A symmetric and diagonal adjacency matrix}

\item{alpha}{The threshold for the maximum number of edges in the partial clique}

\item{c}{The size of the partial clique}
}
\value{
Candidate rows that may be a part of the partial clique
}
\description{
Prunes columns that can be a part of a partial clique of size c
}
