% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{execute_simulation}
\alias{execute_simulation}
\title{This function runs the simulation.}
\usage{
execute_simulation(
  trials_per_level,
  n_nodes,
  clique_fraction,
  alpha,
  seed = 452129
)
}
\arguments{
\item{trials_per_level}{A natural number specifying how many times to run each level}

\item{n_nodes}{A vector of the number of nodes to create a graph with, e.g. c(10) or c(10, 20, 30)}

\item{clique_fraction}{A vector of the clique fraction}

\item{alpha}{A vector of alpha values}

\item{seed}{For setting the seed}
}
\value{
A data frame with the simulation results
}
\description{
This function runs the simulation.
}
