% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a random adjacency matrix with a partial clique}
\usage{
generate_partial_clique(
  n = 10,
  clique_fraction = 0.5,
  clique_edge_density = 0.9
)
}
\arguments{
\item{n}{Positive integer. Number of nodes in the graph.}

\item{clique_fraction}{Numeric in \link{0,1}. Fraction of nodes that form the partial clique.}

\item{clique_edge_density}{Numeric in \link{0,1}. Edge density among the clique nodes.}
}
\value{
A list with element \code{adj_mat}, a symmetric 0/1 matrix with 1s on the diagonal.
}
\description{
Generate a random adjacency matrix with a partial clique
}
