% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique7.R
\name{is_partial_clique_7}
\alias{is_partial_clique_7}
\title{Helper function to check if a set of nodes forms a partial clique}
\usage{
is_partial_clique_7(submatrix, alpha = 1)
}
\arguments{
\item{submatrix}{a submatrix formed by a set of nodes}

\item{alpha}{the minimum required edge density of the partial clique}
}
\value{
TRUE or FALSE depending on whether the given sub-matrix forms a partial clique

TRUE or FALSE depending on whether the given submatrix forms a partial clique
}
\description{
Helper function to check if a set of nodes forms a partial clique
}
