% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW4_simulation_execute.R
\name{my_function}
\alias{my_function}
\title{Title}
\usage{
my_function(num_trials = 5, imp_numbers = 1:25)
}
\arguments{
\item{num_trials}{is Number of Trials per implementation}

\item{imp_numbers}{The Implementation Number (or numbers) to use for the simulation. You can have all by using 1:25}
}
\value{
a list
}
\description{
Title
}
