% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_clique_simulation.R
\name{partial_clique_simulation}
\alias{partial_clique_simulation}
\title{Compute Maximal Partial Clique Simulation}
\usage{
partial_clique_simulation(num_func, num_trials)
}
\arguments{
\item{num_func}{number of functions tested}

\item{num_trials}{number of trials testing}
}
\value{
list of results for each alpha and trial
}
\description{
Compute Maximal Partial Clique Simulation
}
