% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{perform_simulation}
\alias{perform_simulation}
\title{Perform simulation}
\usage{
perform_simulation(imp_numbers, trials, n_vec, alpha_vec)
}
\arguments{
\item{imp_numbers}{List. List of implementations to test}

\item{trials}{Integer. Number of Trials.}

\item{n_vec}{List. n values to test}

\item{alpha_vec}{List. alpha values to test}
}
\value{
A list with simulation results
}
\description{
Perform simulation
}
\examples{
# Run a simulation across multiple graph sizes and alpha thresholds
n_vec <- c(10)
alpha_vec <- c(0.5)

n_level_list <- perform_simulation(
  imp_numbers = 1:10,
  trials = 3,
  n_vec = n_vec,
  alpha_vec = alpha_vec
)

# Record metadata: date of run and session info
date_of_run <- Sys.time()
session_info <- devtools::session_info()

# Save the results to file for reproducibility
save(
  n_level_list, alpha_vec, n_vec,
  date_of_run, session_info,
  file = "~/HW4_simulation.RData"
)

}
