% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_clique.R
\name{reorder_clique}
\alias{reorder_clique}
\title{Reorder adjacency matrix}
\usage{
reorder_clique(adj_mat, order)
}
\arguments{
\item{adj_mat}{a symmetric adjacency matrix}

\item{order}{a permutation of the row and column indices}
}
\value{
a symmetric adjacency matrix with rows and columns reordered according to the given order
}
\description{
Reorder the adjacency matrix according to the given order.
}
\examples{
adj_mat <- generate_partial_clique(10,0.5,0.5)$adj_mat
order <- sample(10,10,replace=FALSE)
reorder_clique(adj_mat, order)
}
