% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run Example: create an example to generate a partial clique}
\usage{
run_example()
}
\value{
A printed message and the result of \code{generate_partial_clique(n = 10, 0.9, 0.8)}.
}
\description{
This function runs an example call to \code{generate_partial_clique()} with
preset parameters. It is primarily intended for demonstration purpose.
}
\examples{
run_example()

}
