% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Compute Maximal Partial Clique Simulation}
\usage{
simulation(num_func, num_trials)
}
\arguments{
\item{num_func}{number of functions tested}

\item{num_trials}{number of trials testing}
}
\value{
results
}
\description{
Compute Maximal Partial Clique Simulation
}
