% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UWBiost561-package.R
\docType{package}
\name{UWBiost561-package}
\alias{UWBiost561}
\alias{UWBiost561-package}
\title{UWBiost561: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Anke Li \email{first.last@example.com}

}
\keyword{internal}
