% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{compute_maximal_partial_clique}
\alias{compute_maximal_partial_clique}
<<<<<<< HEAD
<<<<<<< HEAD
\title{Find the largest partial clique within a given adjacency matrix}
=======
\title{Find the largest partial clique within a given an adjacency matrix}
>>>>>>> 5fcb456f678efdbb507bc1d0586e00ed731c583f
=======
\title{Find the largest partial clique within a given an adjacency matrix}
>>>>>>> 5fcb456f678efdbb507bc1d0586e00ed731c583f
\usage{
compute_maximal_partial_clique(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{an adjacency matrix where within the largest partial clique will be searched}

\item{alpha}{the minimum required edge density of the partial clique}
}
\value{
a list containing indices of the nodes within the clique and the actual edge density of the return partial clique
}
\description{
<<<<<<< HEAD
<<<<<<< HEAD
Find the largest partial clique within a given adjacency matrix
=======
Find the largest partial clique within a given an adjacency matrix
>>>>>>> 5fcb456f678efdbb507bc1d0586e00ed731c583f
=======
Find the largest partial clique within a given an adjacency matrix
>>>>>>> 5fcb456f678efdbb507bc1d0586e00ed731c583f
}
