% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{compute_maximal_partial_clique}
\alias{compute_maximal_partial_clique}
\title{Compute the maximal partial clique in an adjacency matrix}
\usage{
compute_maximal_partial_clique(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{a symmetric adjacency matrix}

\item{alpha}{the edge density of the adjacency matrix}
}
\value{
a list containing the indices of the maximal partial clique, the edge density of the maximal partial clique, and all partial cliques in the adjacency matrix
}
\description{
Enumerate over each row to find where the edges are
Start with the first row as the first partial clique
For each subsequent row, check if it connects with any of the existing partial cliques.
If it does, append to the existing partial cliques; if it doesn't, create a new partial clique for this row.
Return the partial clique with the most nodes
}
\examples{
adj_mat <- generate_partial_clique(10,0.5,0.5)$adj_mat
compute_maximal_partial_clique(adj_mat, 0.5)$clique_idx
}
