% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{compute_maximal_partial_clique}
\alias{compute_maximal_partial_clique}
\title{Compute the maximal partial clique}
\usage{
compute_maximal_partial_clique(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{a symmetric adjacency matrix consisting of only 0 and 1 with 1's along the diagonal and between 5 and 50 rows/columns}

\item{alpha}{edge density}
}
\description{
Computes the maximal partial clique of a given adjacency matrix
}
