% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{compute_maximal_partial_clique}
\alias{compute_maximal_partial_clique}
\title{Compute maximal partial clique}
\usage{
compute_maximal_partial_clique(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{adj_mat Numeric matrix (n×n) of 0/1’s, symmetric, 1’s on diagonal,}

\item{alpha}{Single numeric ∈ \link{0.5, 1}, the minimum required edge‐density.}
}
\value{
A list with clique_idx and edge_density
}
\description{
Compute maximal partial clique
}
