% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{compute_maximal_partial_clique}
\alias{compute_maximal_partial_clique}
\title{find maximal partial clique}
\usage{
compute_maximal_partial_clique(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{the matrix to look for maximal partial clique}

\item{alpha}{density that the partial clique needs to reach}
}
\value{
returns a list of two things, the nodes in the maximal partial clique, and its density
}
\description{
find maximal partial clique
}
