% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{compute_maximal_partial_clique}
\alias{compute_maximal_partial_clique}
\title{Title}
\usage{
compute_maximal_partial_clique(adj_mat, alpha)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
