% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{compute_maximal_partial_clique}
\alias{compute_maximal_partial_clique}
\title{Compute maximal partial clique
This function computes the maximal partial clique by recursively removing the node with the least number of edges until the desired density (alpha) is obtained}
\usage{
compute_maximal_partial_clique(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{An adjacency matrix with edges between a node and itself. Should contain at least 5 nodes and no more than 50.}

\item{alpha}{The density required to be a partial clique, excluding self-edges.}
}
\value{
a list containing the index of the clique \code{clique_idx} and density of the clique
}
\description{
Compute maximal partial clique
This function computes the maximal partial clique by recursively removing the node with the least number of edges until the desired density (alpha) is obtained
}
\examples{
library(UWBiost561)
adj_mat <- diag(1, 10, 10)
adj_mat[1:3, 1:3] <- 1
compute_maximal_partial_clique(adj_mat,
                               alpha = 1)

}
