% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique1.R
\name{compute_maximal_partial_clique1}
\alias{compute_maximal_partial_clique1}
\title{Compute Maximal Partial Clique}
\usage{
compute_maximal_partial_clique1(adj_mat, alpha, verbose = FALSE)
}
\arguments{
\item{adj_mat}{A symmetric 0/1 adjacency matrix with 1s on the diagonal.}

\item{alpha}{A numeric threshold between 0.5 and 1 indicating required edge density.}

\item{verbose}{Logical; whether to print debug output. Default is FALSE.}
}
\value{
A list with:
\describe{
\item{clique_idx}{Vector of node indices in the maximal partial clique.}
\item{edge_density}{Density of the subgraph induced by clique_idx.}
}
}
\description{
Compute Maximal Partial Clique
}
\examples{
n <- 10
clique_fraction <- 0.5
clique_edge_density <- 0.5
alpha <- 0.9
#generate partial clique
graph <- generate_partial_clique(n, clique_fraction, clique_edge_density, seed = NULL)
adj_mat <- graph$adj_mat
#compute maximal partial clique
compute_maximal_partial_clique1(adj_mat, alpha, verbose = FALSE)

}
