% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique10.R
\name{compute_maximal_partial_clique10}
\alias{compute_maximal_partial_clique10}
\title{Compute the maximal partial clique function}
\usage{
compute_maximal_partial_clique10(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{An adjacency matrix representing the connections between nodes
in the graph. The matrix should be symmetric, with ones along the diagonal
representing self-loops and zeros or ones elsewhere representing absence or
presence of edges between nodes, respectively.}

\item{alpha}{The required edge density among the nodes in the maximal partial
clique. This value should be a numeric scalar between 0.5 and 1, inclusive.}
}
\value{
A list containing the index numbers of the nodes forming the maximal
partial clique (\code{clique_idx}) and the corresponding edge density
(\code{edge_density}) among these nodes.
}
\description{
This function computes the largest partial clique given an adjacency matrix
\code{adj_mat} and a required edge density \code{alpha}.
}
\details{
The function iteratively searches for the largest partial clique in the graph
by removing nodes that do not meet the required edge density criterion.
It initializes the partial clique with all nodes in the graph and gradually
removes nodes until the remaining subset of nodes forms a partial clique with
an edge density greater than or equal to the specified \code{alpha}.
The resulting partial clique is returned along with its calculated edge density.
}
