% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique11.R
\name{compute_maximal_partial_clique11}
\alias{compute_maximal_partial_clique11}
\title{Compute partial clique}
\usage{
compute_maximal_partial_clique11(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{an adjacency matrix that is 0,1}

\item{alpha}{minimum edge density}
}
\value{
a list containing the maximum partial clique and its density
}
\description{
Randomly try a bunch of cliques and return the largest one that passes the threshold
}
