% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique12.R
\name{compute_maximal_partial_clique12}
\alias{compute_maximal_partial_clique12}
\title{Title}
\usage{
compute_maximal_partial_clique12(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{matrix}

\item{alpha}{value}
}
\value{
list
}
\description{
Title
}
