% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique13.R
\name{compute_maximal_partial_clique13}
\alias{compute_maximal_partial_clique13}
\title{Title}
\usage{
compute_maximal_partial_clique13(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{adjacency matrix}

\item{alpha}{edge density}
}
\value{
a list
}
\description{
Title
}
