% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique14.R
\name{compute_maximal_partial_clique14}
\alias{compute_maximal_partial_clique14}
\title{Compute the Maximal Partial Clique}
\usage{
compute_maximal_partial_clique14(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{The adjacency matrix representing the graph.}

\item{alpha}{The minimum required edge density for the partial clique, between 0.5 and 1.}
}
\value{
A list containing \code{clique_idx} (numeric vector of indices) and \code{edge_density} (numeric proportion of edges in the partial clique).
}
\description{
Compute the Maximal Partial Clique
}
