% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique2.R
\name{compute_maximal_partial_clique2}
\alias{compute_maximal_partial_clique2}
\title{Compute the largest partial clique given an adjacency matrix
and a required edge density}
\usage{
compute_maximal_partial_clique2(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{an adjacency matrix with only values 0 or 1}

\item{alpha}{required edge density}
}
\value{
a list of index numbers corresponding to the nodes
of maximal partial clique
}
\description{
Compute the largest partial clique given an adjacency matrix
and a required edge density
}
