% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique4.R
\name{compute_maximal_partial_clique4}
\alias{compute_maximal_partial_clique4}
\title{Compute maximal partial clique}
\usage{
compute_maximal_partial_clique4(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{adjacency matrix - symmetric matrix of 0s and 1s with 1s along the diagonal and between 5 to 50 rows/columns}

\item{alpha}{edge density}
}
\value{
vector with index numbers of nodes in maximal partial clique along with the edge density
}
\description{
Computes the maximal partial clique of a given adjacency matrix
}
