% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique5.R
\name{compute_maximal_partial_clique5}
\alias{compute_maximal_partial_clique5}
\title{Compute maximal partial clique}
\usage{
compute_maximal_partial_clique5(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{An adjacency matrix}

\item{alpha}{A required edge density}
}
\value{
A numeric vector of index numbers of the the maximum partial clique
}
\description{
This function first computes the maximal full clique and records the index numbers,
then it tries adding an additional node into the maximal full clique.
If there is one node that makes the new clique satisfy the requirement,
the function tries adding 2 additional nodes to the previous maximal full clique
and repeats the procedure to test different nodes combinations.
The procedure goes on until in the case where it's adding y nodes, no combination of y nodes could satisfy the requirement.
Then it returns the index numbers of the maximal full clique as well as y-1 nodes selected previously as clique_idx.
}
