% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique6.R
\name{compute_maximal_partial_clique6}
\alias{compute_maximal_partial_clique6}
\title{Maximal Partial Clique Computation}
\usage{
compute_maximal_partial_clique6(adj_mat, alpha, ...)
}
\arguments{
\item{adj_mat}{A symmetric adjacency matrix with only values 0 or 1. It must have 1’s along its diagonal, no row- or column-names, and between 5 to 50 rows/columns (inclusive).}

\item{alpha}{A numeric value between 0.5 and 1 (inclusive) specifying the density requirement for the maximal partial clique.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A list with named elements:
\item{clique_idx}{A numeric vector of index numbers corresponding to the nodes in the maximum partial clique.}
\item{edge_density}{The percentage of edges in adj_mat among the nodes in clique_idx.}
}
\description{
This function computes the maximal partial clique with a given adjacency matrix and a edge density threshold.
It iterates through the nodes to identify the largest subset where each pair of nodes is densely connected enough.
}
