% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique7.R
\name{compute_maximal_partial_clique7}
\alias{compute_maximal_partial_clique7}
\title{Find the largest partial clique within a given adjacency matrix}
\usage{
compute_maximal_partial_clique7(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{an adjacency matrix where within the largest partial clique will be searched}

\item{alpha}{the minimum required edge density of the partial clique}
}
\value{
a list containing indices of the nodes within the clique and the actual edge density of the return partial clique
}
\description{
Find the largest partial clique within a given adjacency matrix
}
