% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique8.R
\name{compute_maximal_partial_clique8}
\alias{compute_maximal_partial_clique8}
\title{Compute Maximal Partial Clique}
\usage{
compute_maximal_partial_clique8(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{adjacency matrix to be tested when loaded into the function}

\item{alpha}{the density level of partial clique requested}
}
\value{
a list with the clique index of the largest partial clique in the adjacency matrix, and its edge density
}
\description{
Given an adjacency matrix adj_mat and a required edge density alpha, the program will
determine if the matrix is diagonal, (meaning the nodes have no connection.)
If so, it will return 1 node with clique density of 1.
If the matrix is not diagonal, there is at least one edge.
The program will list the first edge it finds as the largest partial clique.
}
