% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a adjacency matrix representing a matrix containing a partial clique}
\usage{
generate_partial_clique(n, clique_fraction = 0.5, clique_edge_density = 0.5)
}
\arguments{
\item{n}{the number of nodes in the network}

\item{clique_fraction}{the fraction of nodes that are in the partial clique}

\item{clique_edge_density}{the density of edges present between nodes in the clique}
}
\value{
an adjacency matrix containing a partial clique
}
\description{
Function that constructs an adjacency matrix for a variable-sized matrix that
contains a partial clique of variable-density#'
}
