% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a random graph with a partial clique}
\usage{
generate_partial_clique(n, clique_fraction, clique_edge_density)
}
\arguments{
\item{n}{Positive integer. Number of nodes in the graph.}

\item{clique_fraction}{Numeric 0, 1. Fraction of nodes in the partial clique.}

\item{clique_edge_density}{A number between 0 and 1 (inclusive).}
}
\value{
A list with element \code{adj_mat}, the symmetric adjacency matrix.
}
\description{
This function generates a symmetric adjacency matrix with a partial clique.
}
