% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate Random Graph}
\usage{
generate_partial_clique(n, clique_fraction, clique_edge_density)
}
\arguments{
\item{n}{the number of nodes requested in the graph (soon to be matrix)

#n= number of nodes in the graph (along the diagonol of the matrix)}

\item{clique_fraction}{#fraction of nodes connected in our partial clique}

\item{clique_edge_density}{edge density means how much of the clique is fully connected}
}
\value{
adj_mat # a newly-created adjacency matrix that fit the specifications given above
}
\description{
Generate Random Graph
}
