% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a partial clique in a matrix}
\usage{
generate_partial_clique(n, clique_fraction = 0.5, clique_edge_density = 1)
}
\arguments{
\item{n}{the size of the matrix}

\item{clique_fraction}{the fraction of the matrix that should be a clique}

\item{clique_edge_density}{the edge density of the clique}
}
\value{
a list containing the matrix with a partial clique, the randomized matrix, and the order of randomization
}
\description{
Generate a symmetric adjacent matrix of the specified size
Replace the upper left corner of the original matrix with 1s by the clique_edge_density to create a partial clique. Symmetrize the matrix.
Randomize the matrix and keep track of the order of randomization
}
