% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a partial clique}
\usage{
generate_partial_clique(n, clique_fraction, clique_edge_density)
}
\arguments{
\item{n}{the number of nodes in the graph}

\item{clique_fraction}{the fraction of nodes that are part of the partial clique}

\item{clique_edge_density}{the edge density among the nodes in the clique.}
}
\value{
a list named adj_mat
}
\description{
Generate a partial clique
}
\examples{
# call the function to generate a 10*10 adjacency matrix
generate_partial_clique(10,0.5,0.5)

}
