% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a random graph with a partial clique}
\usage{
generate_partial_clique(n, clique_fraction, clique_edge_density)
}
\arguments{
\item{n}{number of nodes in the graph}

\item{clique_fraction}{the fraction of nodes (of the n nodes) that are part of the partial clique}

\item{clique_edge_density}{the edge density among the nodes in the clique}
}
\value{
a random adjacency matrix with a partial clique - it is a symmetric matrix with only values 0 or 1 and has all 1's along its diagonal
}
\description{
Generate a random adjacency matrix with a large partial clique.
}
