% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a Random Partial Clique Adjacency Matrix}
\usage{
generate_partial_clique(
  n,
  clique_fraction,
  clique_edge_density,
  edge_density = 0.2
)
}
\arguments{
\item{n}{A positive integer. Total number of nodes in the graph.}

\item{clique_fraction}{A numeric value between 0 and 1 inclusive. The fraction of nodes that should form a partial clique.}

\item{clique_edge_density}{A numeric value between 0 and 1 inclusive. The density of edges among nodes in the partial clique.}

\item{edge_density}{Single numeric value between 0 and 1, Total edge density use to generate whole matrix, default to 0.2 .}
}
\value{
A list with the following elements:
\describe{
\item{adj_mat}{n x n symmetric adjacency matrix with 0s and 1s, 1s on the diagonal, and no dimnames.}
\item{clique_nodes}{An integer vector indicating which nodes were selected as part of the partial clique.}
\item{instructions}{A message describing how the partial clique was constructed.}
}
}
\description{
This function creates an adjacency matrix representing a graph on \code{n} nodes,
where a subset of nodes form a partial clique. The number of nodes in the clique is
determined by \code{clique_fraction}, and the density of edges among the clique
nodes is controlled by \code{clique_edge_density}.
}
