% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{generate a partial clique}
\usage{
generate_partial_clique(n, clique_fraction, clique_edge_density)
}
\arguments{
\item{n}{number of nodes in the graph}

\item{clique_fraction}{fraction of the n nodes that are part of the partial clique}

\item{clique_edge_density}{the edge density among the nodes in the clique}
}
\value{
a random adjacency matrix that you construct with the partial clique
}
\description{
generate a partial clique
}
