% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Title}
\usage{
generate_partial_clique(n, clique_fraction, clique_edge_density)
}
\arguments{
\item{n}{integer dimension of the square matrix generated by this function}

\item{clique_fraction}{numeric proportion of the aforementioned matrix that is a clique}

\item{clique_edge_density}{numeric proportion of edges expected in the generated matrix's clique}
}
\description{
Title
}
