% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a Random Graph with a Partial Clique}
\usage{
generate_partial_clique(
  n = 50,
  clique_fraction = 0.2,
  clique_edge_density = 1,
  density_low = 0.1
)
}
\arguments{
\item{n}{A positive integer.
It means the number of nodes in the graph, and
therefore the number of rows and columns in the output adjacency matrix.}

\item{clique_fraction}{A single numeric value between 0 and 1 (inclusive).
It is the fraction of the \code{n} nodes that are included in the partial clique.}

\item{clique_edge_density}{A single numeric value between 0 and 1 (inclusive).
It is the desired edge density among the clique nodes.}

\item{density_low}{A single numeric value between 0 and 1 (inclusive).
It is the background probability of an edge in the random graph before the partial
clique structure is imposed. Default value is 0.1.}
}
\value{
A list whose first named element is \code{adj_mat}. \code{adj_mat} is an \verb{n x n} symmetric adjacency
matrix containing only 0s and 1s, with 1s on the diagonal and no row or column names.
}
\description{
\code{generate_partial_clique()} creates an \verb{n x n} symmetric adjacency matrix
for an undirected graph. A subset of nodes is chosen to form a partial clique,
where the number of clique nodes is determined by \code{clique_fraction} and the
number of edges among those clique nodes is determined by
\code{clique_edge_density}.
}
\examples{
output <- generate_partial_clique(
  n = 10,
  clique_fraction = 0.4,
  clique_edge_density = 0.8
)

}
