% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a Random Graph with a Partial Clique}
\usage{
generate_partial_clique(
  n = 10,
  clique_fraction = 0.3,
  clique_edge_density = 0.5
)
}
\arguments{
\item{n}{Integer, the number of nodes in the graph. Defaults to 10.}

\item{clique_fraction}{Numeric, the fraction of nodes that are part of the partial clique. Must be between 0 and 1. Defaults to 0.3.}

\item{clique_edge_density}{Numeric, the density of edges within the clique. Must be between 0 and 1. Defaults to 0.5.}
}
\value{
A list containing a single element \code{adj_mat}, which is an n x n adjacency matrix of the graph. The matrix includes a partial clique according to the specified edge density.
}
\description{
This function creates a random graph containing a specified number of nodes and embeds a partial clique with defined edge density. The graph is represented as an adjacency matrix.
}
