% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_partial_clique.R
\name{generate_partial_clique}
\alias{generate_partial_clique}
\title{Generate a Random Graph with a Partial Clique}
\usage{
generate_partial_clique(n, clique_fraction, clique_edge_density)
}
\arguments{
\item{n}{Integer, the number of nodes in the graph.}

\item{clique_fraction}{Numeric, the fraction of nodes that are part of the partial clique.}

\item{clique_edge_density}{Numeric, the density of edges within the partial clique.}
}
\value{
A list containing the adjacency matrix \code{adj_mat} which has the specified partial clique.
The matrix is symmetric, with zeroes and ones, and ones on its diagonal.
}
\description{
Generate a Random Graph with a Partial Clique
}
