% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique.R
\name{is_partial_clique}
\alias{is_partial_clique}
\title{Helper function to check if a set of nodes forms a partial clique}
\usage{
is_partial_clique(submatrix, alpha = 1)
}
\arguments{
<<<<<<< HEAD
<<<<<<< HEAD
\item{submatrix}{a sub-matrix formed by a set of nodes}
=======
\item{submatrix}{a submatrix formed by a set of nodes}
>>>>>>> 5fcb456f678efdbb507bc1d0586e00ed731c583f
=======
\item{submatrix}{a submatrix formed by a set of nodes}
>>>>>>> 5fcb456f678efdbb507bc1d0586e00ed731c583f

\item{alpha}{the minimum required edge density of the partial clique}
}
\value{
<<<<<<< HEAD
<<<<<<< HEAD
TRUE or FALSE depending on whether the given sub-matrix forms a partial clique
=======
TRUE or FALSE depending on whether the given submatrix forms a partial clique
>>>>>>> 5fcb456f678efdbb507bc1d0586e00ed731c583f
=======
TRUE or FALSE depending on whether the given submatrix forms a partial clique
>>>>>>> 5fcb456f678efdbb507bc1d0586e00ed731c583f
}
\description{
Helper function to check if a set of nodes forms a partial clique
}
