% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{run_simulation}
\alias{run_simulation}
\title{run a simulation study on given functions that computes maximal partial clique for a given matrix}
\usage{
run_simulation(
  trials,
  clique_edge_density_vec,
  clique_fraction = 0.9,
  n = 10,
  alpha = 0.7,
  time_limit = 2,
  imp_numbers = 1:15
)
}
\arguments{
\item{trials}{number of trials to run for each setting each function}

\item{clique_edge_density_vec}{list of edge density among the nodes in the clique of the generated matrix}

\item{clique_fraction}{fraction of the n nodes that are part of the partial clique of the generated matrix}

\item{n}{number of nodes in the graph of the generated matrix}

\item{alpha}{alpha level that the partial clique has to achieve}

\item{time_limit}{maximum time in seconds each implementation can run for.}

\item{imp_numbers}{which implementations to run simulation upon. Need to be between 1 and 15.}
}
\value{
a random adjacency matrix that you construct with the partial clique
}
\description{
run a simulation study on given functions that computes maximal partial clique for a given matrix
}
