% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run Simulation Study for Maximal Partial Clique Implementations}
\usage{
run_simulation(levels_n, num_trials, alpha_vec, time_limit = 15)
}
\arguments{
\item{levels_n}{A vector of integers representing different sizes of the graph (number of nodes).}

\item{num_trials}{An integer representing the number of trials to run for each combination of graph size and edge density.}

\item{alpha_vec}{A vector of numeric values representing different edge densities for the partial clique.}

\item{time_limit}{A numeric value representing the maximum time allowed for each implementation to run (in seconds). Default is 15.}
}
\value{
A list containing the results of the simulation study (\code{level_trial_list}), the alpha values used (\code{alpha_vec}), the date of the run (\code{date_of_run}), and the session information (\code{session_info}).
}
\description{
This function runs a simulation study to evaluate the performance of 25 different implementations of the compute_maximal_partial_clique function.
}
\details{
The function generates random adjacency matrices using the generate_partial_clique function, runs each of the 25 implementations of compute_maximal_partial_clique on these matrices, and stores the results.
}
