% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Title}
\usage{
simulation(density_vec, trials, imp_numbers)
}
\arguments{
\item{density_vec}{a vector of partial clique densities}

\item{trials}{a vector of trials that we want to run}

\item{imp_numbers}{a vector of implementation numbers that we want to simulate over}
}
\value{
a list of trials for different partial clique densities, inside each list is a list of results for each implementation
}
\description{
Title
}
