% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Title}
\usage{
simulation(n, alphas)
}
\arguments{
\item{n}{size of adjacency matrices to be simulated, must be integer greater than or equal to 1}

\item{alphas}{density of adjacency matrices}
}
\value{
results of simulation, 0 for failure to find clique, 1 for success, and NA for function failure
}
\description{
Title
}
